+++ src/main/java/com/shop/tostring/domain/entity/product/ProductEntity.java
@@ -23,7 +23,7 @@ public class ProductEntity extends BaseEntity {
     private int pno;		            // 제품 번호
     private String pname;				// 제품 이름
     private int pprice;					// 제품 가격
-    private int pdiscount;				// 제품 할인율
+    private float pdiscount;				// 제품 할인율
     private int pactive; 				// 제품 상품 상태 : 0 준비중 등
     private String pimg;				// 제품 썸네일
